Conflicts:
	hash-iop.h
	hash-md2.h
	hash-md4.h
	hash-md5.h
	hash-sha1.h
	hash-sha2.h
	hash-sha4.h

Nicolas Pauss (2):
      1699c38dba lib-common: fix compilation with GCC 10.
      6c709c9cf6 lib-common: fix iopc runtime compile error with GCC 10.

Xavier Chantry (1):
      fa0f66c9e7 fix build with clang 10

+++ hash-iop.h
@@ -19,7 +19,7 @@
 struct iop_struct_t;
 
 typedef void (iop_hash_f)(void * nonnull ctx, const void * nonnull input,
-                          int ilen);
+                          ssize_t ilen);
 
 enum {
     IOP_HASH_SKIP_MISSING = 1 << 0, /* Skip missing optional fields         */
